Convert cConsoleLogger to a non-static class
This ensures that Log and LogMessage are only called if the class
initialization succeeded.

Throw an exception if AllocConsole or GetStdHandle fail.
Changed the console writing code to use snprintf and WriteFile.

Don't call CloseHandle of FreeConsole in the destructor, by the time it
is called the OS will have already cleaned up those resources.
This was causing the debugger to show an error as the process exited.